requirements.txt
lark==1.2.2
python-hcl2==5.1.1
PyYAML==6.0.2
ansible==11.0.0
